<refactor>(global): <move cors, logging and recover to package middleware>
